+++ NEWS.md
@@ -5,6 +5,7 @@
 # Version 0.2.1
 - Update for medfate version 0.2.1
 - Revision of runoff
+- Grid simulations with MeteorologyInterpolationData as input
 
 # Version 0.2.0
 - SFI extract functions moved from 'medfate'